Move key vault classes to tessera-app package
