+++ tessera-app/src/main/resources/tessera-spring.xml
@@ -137,12 +137,12 @@
 
     <bean id="keyConfig" factory-bean="config" factory-method="getKeys" />
 
-    <bean id="keyVaultClient" class="com.quorum.tessera.config.keys.vault.KeyVaultAuthenticator" factory-method="getAuthenticatedClient"/>
+    <bean id="keyVaultClient" class="com.quorum.tessera.key.vault.KeyVaultAuthenticator" factory-method="getAuthenticatedClient"/>
 
-    <bean id="keyVaultService" class="com.quorum.tessera.config.keys.vault.KeyVaultService">
+    <bean id="keyVaultService" class="com.quorum.tessera.key.vault.KeyVaultService">
         <constructor-arg ref="keyConfig"/>
         <constructor-arg>
-            <bean id="keyVaultClientDelegate" class="com.quorum.tessera.config.keys.vault.KeyVaultClientDelegate">
+            <bean id="keyVaultClientDelegate" class="com.quorum.tessera.key.vault.KeyVaultClientDelegate">
                 <constructor-arg ref="keyVaultClient"/>
             </bean>
         </constructor-arg>
