+++ tessera-app/src/test/java/com/quorum/tessera/key/KeyPairFactoryTest.java
@@ -1,7 +1,7 @@
 package com.quorum.tessera.key;
 
 import com.quorum.tessera.config.KeyData;
-import com.quorum.tessera.config.keys.vault.KeyVaultService;
+import com.quorum.tessera.key.vault.KeyVaultService;
 import com.quorum.tessera.nacl.Key;
 import com.quorum.tessera.nacl.KeyPair;
 import org.junit.Before;
